CBT : CetTest Whitespace/Indent/Style cleanup
To help in adding cet_test et al to MiniCet, tidy up module to
conform with modern CMake style. No major functionality changes,
though install of tests is commented out as this is likely uneeded
(if tests *are* to be installed, test author should arrange this),
and paths are UPS locked.

Import RunAndCompare module as whilst it is not used in cetlib,
it is a requirement of CetTest.

Add comments throughout cet_test to identifiy problem areas or
those that might be simplified.