Add PATH to the test environment for shell scripts
Shell script wrappers for tests assume the wrapped executable will
be in the PATH. That's set when running under a full UPS build
environment, but not otherwise. PATH assumption also is not completely
robust, and also prevents use of multimode tools.

For shell script wrappers, add PATH to the testing environment directly
using a generator expression to determine the directory holding the
wrapped executable. This increases the robustness of locating the
correct exe, and the generator allows multimode tools to be used
as the path will expand correctly to the used mode.