Create default build type for single mode gens
Use RelWithDebInfo as default build type for single mode
generators. On set default when no build mode is supplied, and
translate CET "Opt" and "Prof" types to the CMake identical types
Release and MinSizerel.

Update assertion management to enable assertions for all build types
if the user has requested it.