Use loadable CET/UPS variable cache when requested
Upstream fix in fnal-cetbuildtools/modern-cmake allows generation
of loadable cmake file in the build dir containing UPS/CET settings.

When building in UPS/CET mode, load this file and fail if it can
not be loaded. Use variable to

- validate package name/version against product_deps value
- set default build mode using CETPKG_TYPE
- Check that compiler matches that required by primary qualifier

Work is still needed to add in install directories, but needs thought
as theey need caching, including the correct docstrings.