Provide UPS/CET policy for install locations
When UPS/CET style build selected, override install locations defined
by GNUInstallDirs with default values following guidelines for UPS
product. Values are forced to cache so they are visible there, but
non-editable following guidelines and original behaviour in
cetbuildtools.

At present, the values set are not derived from those set in the
cetpkg_variable_report.cmake file. This can be done either in the
cmake step, or by writing the values into the file. When the file
is generated, values like the product and flavor qualified directory
are known.