Report-only test demangling of std::string
Use of cetlib::demangle_symbol in fhicl-cpp revealed that whilst
clang/Xcode provide a working cxxabi header/api, it does not
demangle std::string to the expected "string" output (expected
by the fhicl-cpp test anyway). On Xcode, the demangled name is
correct, but resolves to the full underlying type, basic_string,
and the inline namespace. This is also the case for GCC 5.3.

Implement a report only test in demangle_t suite to report typeid
and demangled name for std::string. This is a placeholder to allow
checks on different platforms and compiler versions.