Use full paths in script tests
Certain tests for cetlib wrap a compiled exe inside a shell script.
These tests fail without UPS/cetbuildtools because cet_test assumes
that PATH is set.

Fix script tests so that an absolute path to the test script is used
to run the test. Set PATH in test environments to prepend location
of wrapped executable. This ensures user does not need to have PATH
set correctly, as well as guaranteeing the exact test only is run.