Remove tautological-compare error for Clang
Comparison of unsigned long with min value of unsigned type is
tautologically true as both are unsigned and thus by definition
have the same min value of zero.

Remove tautological less-than comparsion in cetlib:ston specialised
on unsigned. Retain other half of logic test of max values as this
is a valid check.